backend/requirements.txt
flask
flask[async]
supabase
hypercorn
python-dotenv
pydantic
pydantic[email]
bcrypt
dataclasses-json
